# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/createnote": {
            "post": {
                "summary": "create note",
                "description": "create note",
                "operationId": "ac1b3ae4b2f77f1b6e72499a395d033a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note created Sucessfully"
                    },
                    "401": {
                        "description": "Invalid token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/getNoteById": {
            "post": {
                "summary": "Read Note",
                "description": " Read Note ",
                "operationId": "113d4a9232db6d7251c0eb0df352bbd5",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note found Sucessfully"
                    },
                    "404": {
                        "description": "Notes not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/getAllNotes": {
            "get": {
                "summary": "read notes",
                "description": "user read notes",
                "operationId": "e245bcac922abd123d138953b1261495",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "Notes shown suucessfully"
                    },
                    "401": {
                        "description": "No note created by this user"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/updateNoteById": {
            "post": {
                "summary": "update note",
                "description": "update user note",
                "operationId": "a11e25c459bae96f3c862f767bfb8430",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "Updated_title",
                                    "Updated_description",
                                    "id"
                                ],
                                "properties": {
                                    "Updated_title": {
                                        "type": "string"
                                    },
                                    "Updated_description": {
                                        "type": "string"
                                    },
                                    "id": {}
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note successfully updated"
                    },
                    "402": {
                        "description": "Notes not found"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/deleteNoteById": {
            "delete": {
                "summary": "delete note",
                "description": "delete user note",
                "operationId": "563ccd06fc5b23af9eea59ab50ff51fc",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note successfully deleted"
                    },
                    "404": {
                        "description": "Notes not found"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    }
}